+++ lib/ts/recipe/thirdparty/types.ts
@@ -109,6 +109,11 @@ export type PreAPIHookContext =
           action: "GET_AUTHORISATION_URL";
           requestInit: RequestInit;
           url: string;
+      }
+    | {
+          action: "THIRD_PARTY_SIGN_IN_UP";
+          requestInit: RequestInit;
+          url: string;
       };
 
 export type OnHandleEventContext = AuthRecipeModuleOnHandleEventContext;